Refactor markets out of helpers.ts
